Introducing McLib

+++ src/main/java/mchorse/blockbuster/client/gui/dashboard/panels/recording_editor/GuiRecordSelector.java
@@ -6,12 +6,12 @@ import java.util.function.Consumer;
 import org.lwjgl.opengl.GL11;
 
 import mchorse.blockbuster.client.gui.dashboard.GuiDashboard;
-import mchorse.blockbuster.client.gui.framework.GuiTooltip;
-import mchorse.blockbuster.client.gui.framework.elements.GuiElement;
-import mchorse.blockbuster.client.gui.utils.ScrollArea;
-import mchorse.blockbuster.client.gui.utils.ScrollArea.ScrollDirection;
 import mchorse.blockbuster.recording.ActionRegistry;
 import mchorse.blockbuster.recording.actions.Action;
+import mchorse.mclib.client.gui.framework.GuiTooltip;
+import mchorse.mclib.client.gui.framework.elements.GuiElement;
+import mchorse.mclib.client.gui.utils.ScrollArea;
+import mchorse.mclib.client.gui.utils.ScrollArea.ScrollDirection;
 import mchorse.metamorph.client.gui.utils.GuiUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Gui;
@@ -225,7 +225,7 @@ public class GuiRecordSelector extends GuiElement
                     for (Action action : actions)
                     {
                         int y = this.scroll.y + j * 20 - this.vertical.scroll;
-                        int color = MathHelper.hsvToRGB((float) (ActionRegistry.getType(action) - 1) / 20F, 1F, 1F);
+                        int color = MathHelper.hsvToRGB((ActionRegistry.getType(action) - 1) / 20F, 1F, 1F);
 
                         Gui.drawRect(x, y, x + h, y + 20, color + 0x88000000);
                         this.font.drawStringWithShadow(String.valueOf(j), x + 6, y + 6, 0xffffff);
@@ -252,7 +252,7 @@ public class GuiRecordSelector extends GuiElement
             int y = mouseY;
 
             Action action = this.panel.record.getAction(this.tick, this.index);
-            int color = MathHelper.hsvToRGB((float) (ActionRegistry.getType(action) - 1) / 20F, 1F, 1F);
+            int color = MathHelper.hsvToRGB((ActionRegistry.getType(action) - 1) / 20F, 1F, 1F);
 
             Gui.drawRect(x, y, x + h, y + 20, color + 0x88000000);
             this.font.drawStringWithShadow(String.valueOf(this.index), x + 6, y + 6, 0xffffff);
